Adding research paper document
This is a draft document in the format of research paper that I wrote a couple of years ago. The document is unfinished and uploaded upon the request of some of my LinkedIn connections. If you have interest or you are looking for a PhD candidate to work with you on RADAR research topics, please feel free to reach out to me at abd.belissaoui@gmail.com